+++ runtime/include/gpu/chpl-gpu-gen-common.h
@@ -59,7 +59,7 @@ __device__ static inline c_nodeid_t chpl_rt_sublocFromLocaleID(chpl_localeID_t l
 }
 
 __device__ static inline void chpl_gen_comm_get(void *addr, c_nodeid_t node,
-  void* raddr, size_t size, int32_t commID, int ln, int32_t fn)
+  void* raddr, size_t size, int32_t commID, int64_t ln, int32_t fn)
 {
   printf("Warning: chpl_gen_comm_get called inside a GPU kernel. This shouldn't have happened.\n");
   // TODO
@@ -69,14 +69,14 @@ __device__ static inline void chpl_gen_comm_get_unordered(void *addr,
                                                           c_nodeid_t node,
                                                           void* raddr, size_t
                                                           size, int32_t commID,
-                                                          int ln, int32_t fn)
+                                                          int64_t ln, int32_t fn)
 {
   printf("Warning: chpl_gen_comm_get_unordered called inside a GPU kernel. This shouldn't have happened.\n");
   // TODO
 }
 
 __device__ static inline void chpl_gen_comm_put(void* addr, c_nodeid_t node,
-  void* raddr, size_t size, int32_t commID, int ln, int32_t fn)
+  void* raddr, size_t size, int32_t commID, int64_t ln, int32_t fn)
 {
   printf("Warning: chpl_gen_comm_put called inside a GPU kernel. This shouldn't have happened.\n");
   // TODO
@@ -86,7 +86,7 @@ __device__ static inline void chpl_gen_comm_put_unordered(void *addr,
                                                           c_nodeid_t node,
                                                           void* raddr, size_t
                                                           size, int32_t commID,
-                                                          int ln, int32_t fn)
+                                                          int64_t ln, int32_t fn)
 {
   printf("Warning: chpl_gen_comm_put_unordered called inside a GPU kernel. This shouldn't have happened.\n");
   // TODO
@@ -96,7 +96,7 @@ __device__ static inline void chpl_gen_comm_getput_unordered(
                                             c_nodeid_t dstnode, void* dstaddr,
                                             c_nodeid_t srcnode, void* srcaddr,
                                             size_t size, int32_t commID,
-                                            int ln, int32_t fn)
+                                            int64_t ln, int32_t fn)
 {
   printf("Warning: chpl_gen_comm_getput_unordered called inside a GPU kernel. This shouldn't have happened.\n");
   // TODO
@@ -207,7 +207,7 @@ __device__ static inline
 void chpl_gen_comm_put_to_subloc(void* addr,
                                  c_nodeid_t dst_node, c_sublocid_t dst_subloc,
                                  void* raddr, size_t size, int32_t commID,
-                                 int ln, int32_t fn) {
+                                 int64_t ln, int32_t fn) {
   printf("Warning: chpl_gen_comm_get_from_subloc called inside a GPU kernel. This shouldn't have happened.\n");
   // TODO
